Why on earth popcount should return a signed value ?

+++ core/silkworm/common/util.hpp
@@ -111,7 +111,7 @@ inline unsigned int popcount_16(uint16_t value) {
 #ifdef _MSC_VER
     return __popcnt16(value);
 #else
-    return __builtin_popcount(value);
+    return static_cast<unsigned int>(__builtin_popcount(value));
 #endif
 }
